Fix sths34pf80_safe_power_down never reaching drdy == 1.
Issue is related to the fact that sths34pf80_odr_safe_set is
called twice: first time to set odr to 0 (power-down) and second
time to restore original value. But the second time the ctrl1 reg,
since it is passed by value, still contains the original odr (not 0)
and the routine takes a wrong path.
Fixes #3

Signed-off-by: Armando Visconti <armando.visconti@st.com>